Extract oidc callback page, rename Root -> routes
