+++ src/main.tsx
@@ -4,7 +4,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 
-import { App } from "./Root.js";
+import { App } from "./routes.js";
 import { DebugProvider } from "./components/debug/debug-mode.js";
 
 ReactDOM.createRoot(document.getElementById("react-app")!).render(
